Add ping before any requests performed
Inside requests module, sockets are used in blocking mode and current
implementation of LWIP of ESP32 does not support timeouts. This leads
to long timeouts when host is unreachable and this blocks work of whole
reader. As workaround quick ping in polling mode  was added before any
requests are performed.